Simplify IdeaCard futher
